Add internal ELB for readonly API Server
Creates internal ELB for kube-apiserver on the readonly port (8082).

Makes it possible to create ZMON checks that only run once for each
cluster instead of multiple times in case of multiple master nodes.

This also makes the internal ELB for `kube-state-metrics` obsolete since
this can now be proxied through the API server.