Disable creating parameters on the etcd stack
This just disables the creation of cloudformation stack parameters for
the etcd stack. This change is needed to make the CLM update the etcd
stack as part of cluster provisioning. Before we had to update the etcd
stacks manually outside of CLM.

The parameters are not needed since the values passed to senza will be
injected into the right place in the template anyway.