Adding memory MaxResourcePolicy to VPAs
Adding memory MaxResourcePolicy to both prometheus and metrics-server
VPAs. This is so that the VPA doesn't scale prometheus and metrics-server
so much that they become unschedulable.

Note that if `promtheus_mem` is set than the vpa can't *scale up* more than
that, so this should be set to the maximum possible value in a cluster

Signed-off-by: Muhammad Muaaz Saleem <muhammad.muaaz.saleem@zalando.de>