Increase Fluentd Buffer Queue Size
This increases the number of chunks that can be queued to be sent to
S3. The [documentation][1] claims that this number is unlimited when
not set, but the default was in fact [recently set][2] to `1`, which
causes a backlog of chunks to build up when there is a larger number
of log files.

[1]: https://docs.fluentd.org/v1.0/articles/buffer-section#buffering-parameters
[2]: https://github.com/fluent/fluentd/pull/2173

Signed-off-by: Christian Berg <berg.christian@gmail.com>
Signed-off-by: njuettner <nick@zalando.de>