Upgrading StackSet Controller to 1.3.8
The StackSet Controller version 1.3.8 introduces improved support for
Ingress to RouteGroup migration. Check the [release notes][0] for more
details.

This commit upgrades the StackSet controller version deployed to
clusters and its dependency used by the e2e tests.
Also, it adds a new config item called
`stackset_ingress_source_switch_ttl`, responsible for defining the value
of the flag [`--ingress-source-switch-ttl`][1] during the controller
execution. It's required to run the StackSet controller e2e tests as it
expects a TTL of 1 minute, but, also, it allow to configure different
TTLs for different clusters. The default value is 5 minutes.

[0]: https://github.com/zalando-incubator/stackset-controller/releases/tag/v1.3.8
[1]: https://github.com/zalando-incubator/stackset-controller/blob/e72232e799e5ff8ba5b51d29dc5ddff153907162/cmd/stackset-controller/main.go#L56