Update ingress-controller and adjust EW range rule
In #3952 we enabled the Skipper [EastWest Range][0] feature using the
predicates `ClientIP` and `SourceFromLast` as [discussed][1]. It
introduced a problem: other skipper deployments running inside clusters
(not the ingress one) could receive traffic and route it back to the
`skipper-ingress` deployment. This traffic would be denied, since the
`X-Forwarded-For` header won't be handled and its value is as of an
external IP.

This commit updates the kube-ingress-aws-controller to the version
`v0.11.20`. This new version brings among, a fix for NLB HealthCheck
timeouts and dependencies updates, the feature to [deny traffic for
internal domains][2] on the ALB level. It allow us to change the skipper
[east-west range][0] feature to block traffic based just on the
`ClientIP` predicate. This commit, therefore, removes the
`SourceFromLast` predicates previously configured.

[0]: https://skipper.docs.zalando.net/tutorials/operations/#east-west-range
[1]: https://github.com/zalando/skipper/issues/1526#issuecomment-756903746
[2]: https://github.com/zalando-incubator/kube-ingress-aws-controller#deny-traffic-for-internal-domains