Remove affinity configs from skipper
Skipper is already running as host network and host port, so we don't
need to rely on affinity to spread the pods. At the same time we
distrusts cluster autoscaler is taking bad decisions when launching news
nodes based on that.

This commit removes the affinity configs currently configured for
skipper.