Update the kube-metrics-adapter
This commit updates the kube-metrics-adapter. This update contains
multiple dependencies updates. The main changes are the [removal of the
check on the HTTP collector for `identifier`
label](https://github.com/zalando-incubator/kube-metrics-adapter/pull/332)
and the [ignore of ContainerResource
metrics](https://github.com/zalando-incubator/kube-metrics-adapter/pull/326).
The latter generates wrong error messages on the HPA using containers
identification for scaling.