Fix kube-metrics-adapter issues with new ZMON TSDB
This commit updates the `kube-metrics-adapter` deployment to its [last
release][0]. From the version we were using the main change is the
changes to ZMON client and fixes for compatibility with the ZMON new
time series database.

As a consequence, we can revert the `zmon_kairosdb_url` config item to
the main ZMON address, instead of the legacy endpoint.

[0]: https://github.com/zalando-incubator/kube-metrics-adapter/releases/tag/v0.1.11