Allow skipper to scale based on ScalingSchedules
Currently we have skipper scaling based on ZMON. Since we are migrating
away from ZMON for time based scaling, we need to move skipper too for
`ClusterScalingSchedules`. This is the first one, for hyped-article
releases. We will have more once SRE department decide on the
distribution of `ScalingSchedules`. We will have to decide which ones
will scale skipper and in which average.

This commit adds a new object metric for skipper's HPA targeting on the
hyped-article-release `ClusterScalingSchedule` object deployed to all
clusters. The clusters' skippers will scale iff the config item
`skipper_scaling_schedule_hyped_articles_target` is defined.