Set pdb max unavailable to 50% in test clusters
PDBs are setting a max unavailable of 1%. It allows disruption to
deployments with more than 10 replicas, which works great for production
workloads. The test clusters do not have large deployments and nodes
could contain pods of deployments with 0-9 replica, what makes it update
slow.

This commit sets the max unavailable used by the pdb controller to 50%
in tests clusters. This value would handle deployments with too few
replicas.