Allow logging-agent to use OIDC AWS credentials
To allow the logging agent pods to access its own role via OIDC we have
to adjust the IAM role with a new assume role policy. This commit
creates it and removes the older policies attached to the worker and
master role. It is under a config `logging_agent_enable_oidc_iam` which
defaults to false.