Uses `podManagementPolicy: Parallel` for skipper-ingress-redis
`Parallel` pod management tells the StatefulSet controller to launch or
terminate all Pods in parallel, and to not wait for Pods to become
Running and Ready or completely terminated prior to launching or
terminating another Pod.

https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#parallel-pod-management

This should speedup Redis re-scaling.

To avoid
```
updates to statefulset spec for fields other than 'replicas', 'template', and 'updateStrategy' are forbidden
```
due to change of `podManagementPolicy` the statefulset is recreated via
pre-apply deletion that runs once.

Also changes selector `matchLabels` to allow template `application` label change in the future.

Signed-off-by: Alexander Yastrebov <alexander.yastrebov@zalando.de>