Fix DENY policy so oversight is still possible
Currently DENY policy makes it impossible to properly govern this bucket. In particular, we cannot see the configuration and neither can Access Analyzer. This makes it difficult to spot issues and have confidence all is good. This change should fix this.

NOTE: This does NOT grant access. One still needs access to be able to be able to log in to the bucket. However, this explicitly denies listing, getting, deleting, and putting objects.