Create config-items to help during K8 migration
For the Kubernetes service migration, we need to use patched docker
images for kube-proxy and kube-apiserver. The patch allow us to change
the service clusterIP for the new IP in the new ServiceCIDR range,
this behaviour is not allowed by default.
Once the Kubernetes Service CIDR migration is completed this commit
should be undone by setting againg the default docker images for
kube-apiserver and kube-proxy
Patch: https://github.bus.zalan.do/teapot/kubernetes-pipeline/pull/71/files

To avoid disruptions during the migration the configuration from
coredns, flannel, and skipper-internal need to be updated with the new
service CIDR range