Add external-dns config-item policy and annotation
This commit adds the config-tiem external_dns_policy that allows
change the syncronazition policy between Kuberntes and AWS Route53,
the main idea is to use it together with the annotation
`external-dns.alpha.kubernetes.io/exclude: true`.

To the serviceCIDR migration  we need to add an temporary ExternalIP to the
Kubernetes services, but we want to keep the Loadbalancer hostname and avoid
external-dns updates the Route53 entry with the ExternalIP.

To achieve that behavior, we need to temporarily set the external_dns_policy to
upsert-only, and add the annotatation `external-dns.alpha.kubernetes.io/exclude: true`
to the Kubernetes service that has the ExternalIP. Then, external-dns should create
and update the Route53, but will not delete AWS Route53 entries.