fix: Remove changes from routesrv
We should have k8s healthchecks in routesrv to pass e2e load
tests. Tests are started with the tip of the dev branch and later the
new configuration is applied. There is a timespan when old ingress
pods consume routes from the new routesev. If healthcheck is false they
will have no healthz endpoint for some time.

Signed-off-by: Aleksandr Ponimaskin <aleksandr.ponimaskin@zalando.de>