feat: Add sandbox controller
Added the deployment manifest for the sandbox-controller component.

A new cluster configuration item, sandbox_controller_enabled, has been introduced to control its deployment. The controller is disabled by default, and setting this config item to "true" is required to enable its deployment in the cluster.

A conditional deletion rule was added to ensure the deployment is removed when the configuration item is not set to "true".